Subway_Python_Component/requirements.txt
supabase==2.16.0
python-dotenv==1.0.1
Flask==3.1.1
requests==2.32.3
pytz==2024.2
Flask-SocketIO==5.5.1